Add ErrorHandler

+++ Calculator/Operations/Add.cs
@@ -11,26 +11,10 @@ namespace Calculator.Operations
     {
         public string Name => "Dodaj";
 
-        public double Calculate(int number1, int number2)      // operacje to mają być delegaty
+        public int Calculate(int number1, int number2)   
         {
             return number1 + number2;
         }
-        
-
-        //int Miltiply(int number1, int number2)
-        //{
-        //    return number1 * number2;
-        //}
-
-        //double Divide(int number1, int number2)
-        //{
-        //    return (double)number1 / number2;
-        //}
-
-        //int Power(int number1, int number2)
-        //{
-        //    return (int)Math.Pow(number1, number2);     // czy lepiej zwrócić double?
-        //}
-
+              
     }
 }
